requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
channels==4.1.0
charset-normalizer==3.3.2
Django==4.2.9
django-bootstrap-v5==1.0.11
django-channels==0.7.0
django-model-signals==0.4.3
django-redis==5.4.0
djangorestframework==3.15.1
idna==3.7
injector==0.21.0
oauthlib==3.2.2
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.4
requests==2.31.0
requests-oauthlib==2.0.0
setuptools==69.5.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.1
